requirements.txt
Django~=5.0
pandas~=2.2
python-docx~=1.1
PyPDF2~=3.0
openpyxl~=3.1
